+++ Catalog/Catalog.cpp
@@ -1457,7 +1457,7 @@ const int Catalog::getColumnIdBySpiUnlocked(const int table_id, const size_t spi
 
   auto spx = spi;
   int phi = 0;
-  if (spx >= SPIMAP_MAGIC1)  // see Catalog.h
+  if (spx >= size_t(SPIMAP_MAGIC1))  // see Catalog.h
   {
     phi = (spx - SPIMAP_MAGIC1) % SPIMAP_MAGIC2;
     spx = (spx - SPIMAP_MAGIC1) / SPIMAP_MAGIC2;
